add negative test for LocationController and fis CategoryService

+++ app/src/main/java/ru/tbank/service/CategoryService.java
@@ -6,7 +6,6 @@ import ru.tbank.model.Category;
 import ru.tbank.repository.CategoryRepository;
 
 import java.util.Collection;
-import java.util.Optional;
 
 @Service
 public class CategoryService {
@@ -21,7 +20,7 @@ public class CategoryService {
     }
 
     public Category getCategoryById(int id) {
-        return Optional.ofNullable(repository.findById(id))
+        return repository.findById(id)
                 .orElseThrow(() -> new CategoryNotFoundException(id));
     }
 
@@ -30,11 +29,16 @@ public class CategoryService {
     }
 
     public Category updateCategory(int id, Category category) {
-        return Optional.ofNullable(repository.update(id, category))
-                .orElseThrow(() -> new CategoryNotFoundException(id));
+        if (!repository.existsById(id)) {
+            throw new CategoryNotFoundException(id);
+        }
+        return repository.update(id, category);
     }
 
     public void deleteCategory(int id) {
+        if (!repository.existsById(id)) {
+            throw new CategoryNotFoundException(id);
+        }
         repository.delete(id);
     }
 }
